+++ src/Symfony/Component/HttpKernel/Controller/ControllerResolverInterface.php
@@ -52,6 +52,8 @@ interface ControllerResolverInterface
      * @return array An array of arguments to pass to the controller
      *
      * @throws \RuntimeException When value for argument given is not provided
+     *
+     * @deprecated This method is deprecated as of 3.1 and will be removed in 4.0. Please use the {@see ArgumentResolverInterface} instead.
      */
     public function getArguments(Request $request, $controller);
 }